Rename godot-wasm.h to wasm.h to reflect class
